WIP Refactoring some of Axis.java functions into TickMarks.Java
