+++ src/main/java/org/jfree/chart/axis/TickMarks.java
@@ -1,0 +1,306 @@
+package org.jfree.chart.axis;
+
+import org.jfree.chart.event.AxisChangeEvent;
+import org.jfree.chart.internal.Args;
+import org.jfree.chart.internal.PaintUtils;
+
+import java.awt.*;
+import java.io.Serializable;
+import java.util.Objects;
+
+public class TickMarks implements Serializable, Cloneable {
+
+    /** The default tick marks visible ({@code true}). */
+    public static final boolean DEFAULT_TICK_MARKS_VISIBLE = true;
+
+    /** The default tick stroke ({@code BasicStroke(0.5f)}). */
+    public static final Stroke DEFAULT_TICK_MARK_STROKE = new BasicStroke(0.5f);
+
+    /** The default tick paint ({@code Color.GRAY}). */
+    public static final Paint DEFAULT_TICK_MARK_PAINT = Color.GRAY;
+
+    /** The default tick mark inside length ({@code 0.0f}). */
+    public static final float DEFAULT_TICK_MARK_INSIDE_LENGTH = 0.0f;
+
+    /** The default tick mark outside length ({@code 2.0f}). */
+    public static final float DEFAULT_TICK_MARK_OUTSIDE_LENGTH = 2.0f;
+
+    /**
+     * A flag that indicates whether or not major tick marks are visible for
+     * the axis.
+     */
+    private boolean tickMarksVisible;
+
+    /**
+     * The length of the major tick mark inside the data area (zero
+     * permitted).
+     */
+    private float tickMarkInsideLength;
+
+    /**
+     * The length of the major tick mark outside the data area (zero
+     * permitted).
+     */
+    private float tickMarkOutsideLength;
+
+    /**
+     * A flag that indicates whether or not minor tick marks are visible for the
+     * axis.
+     */
+    private boolean minorTickMarksVisible;
+
+    /**
+     * The length of the minor tick mark inside the data area (zero permitted).
+     */
+    private float minorTickMarkInsideLength;
+
+    /**
+     * The length of the minor tick mark outside the data area (zero permitted).
+     */
+    private float minorTickMarkOutsideLength;
+
+    /** The stroke used to draw tick marks. */
+    private transient Stroke tickMarkStroke;
+
+    /** The paint used to draw tick marks. */
+    private transient Paint tickMarkPaint;
+
+    /**
+     * Returns the flag that indicates whether or not the minor tick marks are
+     * showing.
+     *
+     * @return The flag that indicates whether or not the minor tick marks are
+     *         showing.
+     *
+     */
+    public boolean isMinorTickMarksVisible() {
+        return this.minorTickMarksVisible;
+    }
+
+    /**
+     * Sets the flag that indicates whether or not the minor tick marks are
+     * showing and sends an {@link AxisChangeEvent} to all registered
+     * listeners.
+     *
+     * @param flag  the flag.
+     *
+     * @see #isMinorTickMarksVisible()
+     */
+    public void setMinorTickMarksVisible(boolean flag, Axis axis) {
+        if (flag != this.minorTickMarksVisible) {
+            this.minorTickMarksVisible = flag;
+            axis.fireChangeEvent();
+        }
+    }
+
+    /**
+     * Returns the flag that indicates whether or not the tick marks are
+     * showing.
+     *
+     * @return The flag that indicates whether or not the tick marks are
+     *         showing.
+     *
+     */
+    public boolean isTickMarksVisible() {
+        return this.tickMarksVisible;
+    }
+
+    /**
+     * Sets the flag that indicates whether or not the tick marks are showing
+     * and sends an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param flag  the flag.
+     *
+     * @see #isTickMarksVisible()
+     */
+    public void setTickMarksVisible(boolean flag, Axis axis) {
+        if (flag != this.tickMarksVisible) {
+            this.tickMarksVisible = flag;
+            axis.fireChangeEvent();
+        }
+    }
+
+    /**
+     * Returns the inside length of the tick marks.
+     *
+     * @return The length.
+     *
+     * @see #getTickMarkOutsideLength()
+     */
+    public float getTickMarkInsideLength() {
+        return this.tickMarkInsideLength;
+    }
+
+    /**
+     * Sets the inside length of the tick marks and sends
+     * an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param length  the new length.
+     *
+     * @see #getTickMarkInsideLength()
+     */
+    public void setTickMarkInsideLength(float length, Axis axis) {
+        this.tickMarkInsideLength = length;
+        axis.fireChangeEvent();
+    }
+
+    /**
+     * Returns the outside length of the tick marks.
+     *
+     * @return The length.
+     *
+     * @see #getTickMarkInsideLength()
+     */
+    public float getTickMarkOutsideLength() {
+        return this.tickMarkOutsideLength;
+    }
+
+    /**
+     * Sets the outside length of the tick marks and sends
+     * an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param length  the new length.
+     *
+     * @see #getTickMarkInsideLength()
+     */
+    public void setTickMarkOutsideLength(float length, Axis axis) {
+        this.tickMarkOutsideLength = length;
+        axis.fireChangeEvent();
+    }
+
+    /**
+     * Returns the stroke used to draw tick marks.
+     *
+     * @return The stroke (never {@code null}).
+     *
+     */
+    public Stroke getTickMarkStroke() {
+        return this.tickMarkStroke;
+    }
+
+    /**
+     * Sets the stroke used to draw tick marks and sends
+     * an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param stroke  the stroke ({@code null} not permitted).
+     *
+     * @see #getTickMarkStroke()
+     */
+    public void setTickMarkStroke(Stroke stroke, Axis axis) {
+        Args.nullNotPermitted(stroke, "stroke");
+        if (!this.tickMarkStroke.equals(stroke)) {
+            this.tickMarkStroke = stroke;
+            axis.fireChangeEvent();
+        }
+    }
+
+    /**
+     * Returns the paint used to draw tick marks (if they are showing).
+     *
+     * @return The paint (never {@code null}).
+     *
+     */
+    public Paint getTickMarkPaint() {
+        return this.tickMarkPaint;
+    }
+
+    /**
+     * Sets the paint used to draw tick marks and sends an
+     * {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param paint  the paint ({@code null} not permitted).
+     *
+     * @see #getTickMarkPaint()
+     */
+    public void setTickMarkPaint(Paint paint, Axis axis) {
+        Args.nullNotPermitted(paint, "paint");
+        this.tickMarkPaint = paint;
+        axis.fireChangeEvent();
+    }
+
+    /**
+     * Returns the inside length of the minor tick marks.
+     *
+     * @return The length.
+     *
+     * @see #getMinorTickMarkOutsideLength()
+     */
+    public float getMinorTickMarkInsideLength() {
+        return this.minorTickMarkInsideLength;
+    }
+
+    /**
+     * Sets the inside length of the minor tick marks and sends
+     * an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param length  the new length.
+     *
+     * @see #getMinorTickMarkInsideLength()
+     */
+    public void setMinorTickMarkInsideLength(float length, Axis axis) {
+        this.minorTickMarkInsideLength = length;
+        axis.fireChangeEvent();
+    }
+
+    /**
+     * Returns the outside length of the minor tick marks.
+     *
+     * @return The length.
+     *
+     * @see #getMinorTickMarkInsideLength()
+     */
+    public float getMinorTickMarkOutsideLength() {
+        return this.minorTickMarkOutsideLength;
+    }
+
+    /**
+     * Sets the outside length of the minor tick marks and sends
+     * an {@link AxisChangeEvent} to all registered listeners.
+     *
+     * @param length  the new length.
+     *
+     * @see #getMinorTickMarkInsideLength()
+     */
+    public void setMinorTickMarkOutsideLength(float length, Axis axis) {
+        this.minorTickMarkOutsideLength = length;
+        axis.fireChangeEvent();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == this) {
+            return true;
+        }
+        if (!(obj instanceof TickMarks)) {
+            return false;
+        }
+        TickMarks that = (TickMarks) obj;
+        if (this.tickMarksVisible != that.tickMarksVisible) {
+            return false;
+        }
+        if (this.tickMarkInsideLength != that.tickMarkInsideLength) {
+            return false;
+        }
+        if (this.tickMarkOutsideLength != that.tickMarkOutsideLength) {
+            return false;
+        }
+        if (!PaintUtils.equal(this.tickMarkPaint, that.tickMarkPaint)) {
+            return false;
+        }
+        if (!Objects.equals(this.tickMarkStroke, that.tickMarkStroke)) {
+            return false;
+        }
+        if (this.minorTickMarksVisible != that.minorTickMarksVisible) {
+            return false;
+        }
+        if (this.minorTickMarkInsideLength != that.minorTickMarkInsideLength) {
+            return false;
+        }
+        if (this.minorTickMarkOutsideLength
+                != that.minorTickMarkOutsideLength) {
+            return false;
+        }
+        return true;
+    }
+
+}
